+++ backend/src/main/java/com/example/backend/Repository/TaskCardRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 import java.util.Optional;
 
-public interface TaskCardRepository extends JpaRepository<TaskCard,Integer> {
+public interface TaskCardRepository extends JpaRepository<TaskCard,Long> {
     Optional<TaskCard> findByName(String name);
 }
